Deepen web research and tool lifecycle modules
Move web research job state, artifact handling, history loading, and manager UI out of the extension entrypoint. Move managed tool availability and active-tool syncing into a dedicated module.

Research artifacts now use Markdown with YAML frontmatter metadata, with history loading kept compatible with legacy heading-based reports. Document the artifact metadata format and add domain context for future architecture reviews.

Verification: npm run check; npm test; npx vitest run test/tool-availability.test.ts. Note: npm run format:check still fails on pre-existing .sc/last-chat-settings.json trailing newline issue.

AI-Assisted-By: OpenAI Codex
